Rename the project to crontainer :sparkles:
